+++ packages/analytics-js-common/src/v1.1/utils/config_to_integration_names.js
@@ -20,6 +20,7 @@ const configToIntNames = {
   FULLSTORY: 'Fullstory',
   TVSQUARED: 'TVSquared',
   GA4: 'GA4',
+  GA4_V2: 'GA4_V2',
   MOENGAGE: 'MoEngage',
   AM: 'Amplitude',
   PENDO: 'Pendo',